Dockerizing-DRF-App/requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
colorama==0.4.1
Django==2.2.4
djangorestframework==3.10.2
importlib-metadata==0.19
more-itertools==7.2.0
packaging==19.1
pluggy==0.12.0
psycopg2==2.8.3
py==1.8.0
pyparsing==2.4.2
pytest==5.0.1
pytest-django==3.5.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
wcwidth==0.1.7
zipp==0.5.2
